backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.4",
    "jinja2>=3.1.6",
    "openai-agents>=0.8.1",
    "openai-chatkit>=1.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
